requirements.txt
alembic==1.7.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.3.5
pip==22.2
protobuf==4.21.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==63.2.0
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
xmltodict==0.13.0
zipp==3.8.1
